refactor: moved Parser\ServerRequest -> Http\Parser
