+++ src/Http/Parser/UriParser.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace BitFrame\Parser\ServerRequest;
+namespace BitFrame\Http\Parser;
 
 use function rtrim;
 use function parse_url;
